Fix for minor issues as reported by cppcheck. Z80_interface had an issue with r_register not being assigned a value by default. In gfx.c, data16 was assigning itself, which is essentially useless. Seems to have been done for alignment in case gfx->negative is true ?? To check.

+++ mednafen/ngp/T6W28_Apu.cpp
@@ -262,7 +262,7 @@ void T6W28_Apu::reset()
 
 void T6W28_Apu::run_until( sms_time_t end_time )
 {
-   int i;
+   uint_fast8_t i;
    require( end_time >= last_time ); // end_time must not be before previous time
 
    if ( end_time > last_time )
